Revert "Added a PDF file for QRCode"
