+++ app/controllers/deals_controller.rb
@@ -86,15 +86,6 @@ class DealsController < ApplicationController
 
   def show
     @qr = RQRCode::QRCode.new(@deal.id.to_s + "_" + @deal.created_at.to_s).to_img.resize(200, 200).to_data_url
-    respond_to do |format|
-      format.html
-      if @deal.redeemable
-        format.pdf do
-          pdf = QrcodePdf.new(@deal)
-          send_data pdf.render, filename: "QRCodes_of_#{@deal.title}.pdf", type: "application/pdf"
-        end
-      end
-    end
   end
 
   def destroy
@@ -116,6 +107,52 @@ class DealsController < ApplicationController
     end
     redirect_to deals_path
   end
+
+  def format_days (deal_day)
+    deal_days = [deal_day.mon, deal_day.tue, deal_day.wed, deal_day.thur, deal_day.fri, deal_day.sat, deal_day.sun ]
+    days = ['Mon', 'Tue', 'Wed', 'Thur', 'Fri', 'Sat', 'Sun']
+    dealperiod = ""
+    i = 0
+
+    while i<7
+      # For first day which is selected, no need to put comma
+      if deal_days[i] && dealperiod == ""
+        dealperiod = dealperiod + days[i]
+        # If the string already contains thing, then need put comma
+      elsif deal_days[i] && dealperiod != ""
+        dealperiod = dealperiod + ", " + days[i]
+      end
+
+      # Inner while loop is to string consecutive days together if the current day is selected
+      if deal_days[i]
+        j = i + 1
+        while j<7
+          # If the day is not selected, break out of loop
+          if !deal_days[j]
+            break
+          elsif j == 6 && deal_days[j]
+            dealperiod = dealperiod + "-" + days[j]
+            i = j+1
+            break
+          elsif j == 6 && !deal_days[j]
+            i = j+1
+            break
+            # If the day is selected, and next one is selected, just continue
+          elsif deal_days[j] && deal_days[j+1]
+            j = j+1
+            # If day is selected, and next one is not, place "- day" and break
+          elsif deal_days[j] && !deal_days[j+1]
+            dealperiod = dealperiod + "-" + days[j]
+            i = j + 1
+            break
+          end
+        end
+      end
+      i = i +1
+    end
+    dealperiod
+  end
+  helper_method :format_days
 
   private
   # Use callbacks to share common setup or constraints between actions.
