+++ app/views/deal_mailer/update_deal_email.html.erb
@@ -22,7 +22,7 @@
 
   <% for deal_day in @deal.deal_days%>
       <br>
-      <strong><%= DealService.format_days(deal_day) %>:</strong>
+      <strong><%= format_days (deal_day) %>:</strong>
       <% for deal_time in deal_day.deal_times%>
           <%=h deal_time.started_at.strftime("%H:%M") %> to <%=h deal_time.ended_at.strftime("%H:%M") %>&nbsp;&nbsp;
           <% end %><% end %>
@@ -31,6 +31,12 @@
 <p>
   <strong><font size = "2">Terms and Condition</font></strong><br>
   <%= @deal.t_c %>
+</p>
+
+
+<p>
+  <strong><font size = "4">Deal Image</font></strong><br>
+  <%= image_tag @deal.image.url, :size => "150x150" %>
 </p>
 
 <p>
@@ -59,7 +65,15 @@
 
 <p>
   <% if @deal.redeemable %>
-      To download your QRCodes, please press the following link <%= @pdf %><br>
+      Showed below are the QRcodes for each venue. Please ensure you download the correct QRcode for each venue.
+      To download the QRcode you can right click on it and choose save image. You can then print the QRcode out and when
+      your customers wish to redeem the deal, just ask your customer to scan the QR code using our customer scanner in our
+      app.<br>
+      <% @deal.venues.each do |venue| %>
+          <strong>QR Code for <%= venue.name %></strong><br>
+          <img src="https://chart.googleapis.com/chart?chs=100x100&amp;cht=qr&amp;chl=<%= @deal.id.to_s + '_' + venue.id.to_s + '_' + @deal.created_at.to_s%>" alt="QR code">
+          <br>
+      <% end %>
   <% end %>
 </p>
