+++ app/views/form/_deal_show_form.erb
@@ -19,11 +19,10 @@
 
   <% for deal_day in @deal.deal_days%>
       <br>
-      <strong><%= DealService.format_days (deal_day) %>:</strong>
+      <strong><%= format_days (deal_day) %>:</strong>
       <% for deal_time in deal_day.deal_times%>
           <%=h deal_time.started_at.strftime("%H:%M") %> to <%=h deal_time.ended_at.strftime("%H:%M") %>&nbsp;&nbsp;
-      <% end %>
-  <% end %>
+      <% end %><% end %>
 </p>
 
 <p>
@@ -71,6 +70,10 @@
 
 <p>
   <% if @deal.redeemable %>
-    <%= link_to "Download QRcodes for your deal here", deal_path(@deal, format: "pdf") %>
+      <% @deal.venues.each do |venue| %>
+          <strong>QR Code for <%= venue.name %></strong><br>
+          <% @qr = RQRCode::QRCode.new(@deal.id.to_s + "_" + venue.id.to_s + "_" + @deal.created_at.to_s).to_img.resize(100, 100).to_data_url %>
+          <span><%= image_tag @qr %><br></span>
+      <% end %>
   <% end %>
 </p>
